requirements.txt
moviepy==1.0.3
openai==1.69.0
python-dotenv==1.0.0
boto3==1.34.69
requests==2.31.0
psutil==5.9.8
Pillow==10.2.0
scipy==1.12.0
numpy==1.26.4
decorator>=4.0.2,<5.0
imageio==2.33.1
imageio-ffmpeg==0.4.9
tqdm==4.66.2
ffmpeg-python==0.2.0
botocore==1.34.69
urllib3==2.2.1
httpx==0.27.0 